Update load-shed.xlsm
Went through some exceltations to split the suburb extensions into different rows which, in theory, should be a more useful unit of analysis for geocoding.
Two problems: some extensions are in two groups, and I can't find a way of geocoding these extensions. Need to find a source for their locations...